backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
PyMuPDF>=1.23.0
openai>=1.12.0
groq>=0.4.0
curl_cffi>=0.6.2
playwright>=1.42.0
python-multipart>=0.0.9
chromadb>=0.4.22
sse-starlette>=1.8.2
httpx>=0.26.0
aiofiles>=23.2.0
